오류 수정 및 데이터 표현 개선
1. 카드번호에서 알파벳 제거
2. 요청패킷의 Volume 소수점 2자리로 수정
3. 소수점이 필요한 항목의 자료형을 decimal로 통일
4. MockPOS, MockPayAPI 문서 오류 및 오타 수정

+++ PayPro.Main/Local/ViewModels/MainWindowViewModel.cs
@@ -357,7 +357,7 @@ namespace PayPro.Main.Local.ViewModels
             var fuelTypes = new[] { "휘발유", "경유" };
             var paymentMethods = new[] { "카드", "간편", "현금" };
             var statuses = new[] { "승인", "실패" };
-            double pricePerLiter = 1652.89; // TODO: 하드코딩 수정
+            decimal pricePerLiter = 1652.89M; // TODO: 하드코딩 수정
 
             var transactions = Enumerable.Range(1, 20).Select(i =>
             {
